+++ stat.go
@@ -2,12 +2,12 @@ package goope
 
 // ValueRange represents an inclusive range of integers [Start, End].
 type ValueRange struct {
-	Start int
-	End   int
+	Start int64
+	End   int64
 }
 
 // NewValueRange creates a new ValueRange with the given start and end values.
-func NewValueRange(start, end int) (ValueRange, error) {
+func NewValueRange(start, end int64) (ValueRange, error) {
 	if start > end {
 		return ValueRange{}, ErrInvalidRangeLimits
 	}
@@ -15,12 +15,12 @@ func NewValueRange(start, end int) (ValueRange, error) {
 }
 
 // Size returns the number of integers in the range.
-func (r ValueRange) Size() int {
+func (r ValueRange) Size() int64 {
 	return r.End - r.Start + 1
 }
 
 // Contains checks if the given value is within the range [Start, End].
-func (r ValueRange) Contains(value int) bool {
+func (r ValueRange) Contains(value int64) bool {
 	return value >= r.Start && value <= r.End
 }
 
@@ -30,7 +30,7 @@ func (r ValueRange) Copy() ValueRange {
 }
 
 // sampleHGD maps a specific output-range position (nsample) back to an input-range value, using the hypergeometric distribution as the sampling method.
-func sampleHGD(inRange, outRange ValueRange, nsample int, coins <-chan bool) (int, error) {
+func sampleHGD(inRange, outRange ValueRange, nsample int64, coins <-chan bool) (int64, error) {
 	inSize := inRange.Size()
 	outSize := outRange.Size()
 
@@ -55,7 +55,7 @@ func sampleHGD(inRange, outRange ValueRange, nsample int, coins <-chan bool) (in
 	return inSample, nil
 }
 
-func floorDiv(a, b int) int {
+func floorDiv(a, b int64) int64 {
 	q := a / b
 	// if the result was negative and there's a remainder, round down instead of up
 	if (a^b) < 0 && a*b != a {
@@ -65,7 +65,7 @@ func floorDiv(a, b int) int {
 }
 
 // sampleUniform picks a value from inRange using binary search guided by bits from coins. Each bit decides whether to take the lower or upper half of the current range, narrowing it until only one value remains.
-func sampleUniform(inRange ValueRange, coins <-chan bool) (int, error) {
+func sampleUniform(inRange ValueRange, coins <-chan bool) (int64, error) {
 	cur := inRange.Copy()
 	if cur.Size() == 0 {
 		return 0, ErrInvalidRanges
